+++ components/diary/DiaryStoryModal.tsx
@@ -14,19 +14,9 @@ import {
 import { motion, AnimatePresence, PanInfo } from 'framer-motion';
 import { useSwipeable } from 'react-swipeable';
 import type { DiaryDetail } from '@/types/diary';
-import type { Comment } from '@/types/comment';
-import {
-  createComment,
-  getRootComments,
-  getReplies,
-  deleteComment,
-  updateComment,
-} from '@/lib/api/comment';
 import { formatTimeAgo, formatYearMonthDayDots } from '@/lib/utils/date';
 import { getServerURL } from '@/lib/utils/url';
-import useAuthStore from '@/components/store/authStore';
-import CommentItem from './CommentItem';
-import CommentInput from './CommentInput';
+import StoryCommentModal from './StoryCommentModal';
 
 interface DiaryStoryModalProps {
   diary: DiaryDetail;
@@ -34,13 +24,6 @@ interface DiaryStoryModalProps {
   onCommentViewToggle?: (isOpen: boolean) => void;
 }
 
-interface CommentRepliesState {
-  list: Comment[];
-  page: number;
-  hasMore: boolean;
-  isLoading: boolean;
-  isShown: boolean;
-}
 
 const formatCount = (count: number): string => {
   if (count >= 1_000_000) {
@@ -58,25 +41,9 @@ const DiaryStoryModal = ({
   onCommentViewToggle,
 }: DiaryStoryModalProps) => {
   const [currentImageIndex, setCurrentImageIndex] = useState(0);
-  const [comments, setComments] = useState<Comment[]>([]);
-  const [commentReplies, setCommentReplies] = useState<
-    Record<number, CommentRepliesState>
-  >({});
-  const [page, setPage] = useState(0);
-  const [hasMore, setHasMore] = useState(true);
-  const [isLoadingComments, setIsLoadingComments] = useState(false);
   const [totalComments, setTotalComments] = useState(diary.commentCount);
-  const [newComment, setNewComment] = useState('');
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [replyTo, setReplyTo] = useState<Comment | null>(null);
-  const [scrollToCommentId, setScrollToCommentId] = useState<number | null>(
-    null,
-  );
-  const [editingComment, setEditingComment] = useState<Comment | null>(null);
   const [isCommentViewOpen, setIsCommentViewOpen] = useState(false);
 
-  const inputRef = useRef<HTMLInputElement | null>(null);
-  const { isLoggedIn, user } = useAuthStore();
   const serverUrl = getServerURL();
   const router = useRouter();
 
@@ -85,312 +52,18 @@ const DiaryStoryModal = ({
     onClose();
   };
 
-
-  const fetchComments = async (isNewFetch: boolean = false) => {
-    if (isLoadingComments || (!hasMore && !isNewFetch)) return;
-    setIsLoadingComments(true);
-    const pageToFetch = isNewFetch ? 0 : page;
-    try {
-      const data = await getRootComments(diary.diaryId, pageToFetch, 10);
-      setComments(prev =>
-        isNewFetch ? data.content : [...prev, ...data.content],
-      );
-      setPage(pageToFetch + 1);
-      setHasMore(!data.last);
-      if (isNewFetch) {
-        setTotalComments(diary.commentCount);
-      }
-    } catch (error) {
-      console.error('댓글을 불러오는데 실패했습니다:', error);
-    } finally {
-      setIsLoadingComments(false);
-    }
+  const handleUpdateCommentCount = (count: number) => {
+    setTotalComments(count);
   };
 
-  const handleToggleReplies = async (comment: Comment) => {
-    const currentState = commentReplies[comment.id] || {
-      isShown: false,
-      list: [],
-      page: 0,
-      hasMore: comment.replyCount > 0,
-      isLoading: false,
-    };
-    if (currentState.isShown) {
-      setCommentReplies(prev => ({
-        ...prev,
-        [comment.id]: { ...currentState, isShown: false },
-      }));
-    } else {
-      setCommentReplies(prev => ({
-        ...prev,
-        [comment.id]: { ...currentState, isShown: true },
-      }));
-      if (currentState.list.length === 0 && currentState.hasMore) {
-        await handleFetchReplies(comment.id);
-      }
-    }
+  const handleCloseCommentModal = () => {
+    setIsCommentViewOpen(false);
   };
-
-  const handleFetchReplies = async (commentId: number) => {
-    const currentState = commentReplies[commentId] || {
-      list: [],
-      page: 0,
-      hasMore: true,
-      isLoading: false,
-      isShown: true,
-    };
-    if (currentState.isLoading || !currentState.hasMore) return;
-
-    setCommentReplies(prev => ({
-      ...prev,
-      [commentId]: { ...currentState, isLoading: true },
-    }));
-
-    try {
-      const data = await getReplies(commentId, currentState.page, 5);
-      setCommentReplies(prev => {
-        const currentReplies = prev[commentId]?.list || [];
-        const existingReplyIds = new Set(currentReplies.map(c => c.id));
-        const newUniqueReplies = data.content.filter(
-          c => !existingReplyIds.has(c.id),
-        );
-
-        return {
-          ...prev,
-          [commentId]: {
-            ...prev[commentId],
-            list: [...currentReplies, ...newUniqueReplies],
-            page: prev[commentId].page + 1,
-            hasMore: !data.last,
-            isLoading: false,
-          },
-        };
-      });
-    } catch (error) {
-      console.error('답글을 불러오는데 실패했습니다:', error);
-      setCommentReplies(prev => ({
-        ...prev,
-        [commentId]: { ...prev[commentId], isLoading: false },
-      }));
-    }
-  };
-
-  useEffect(() => {
-    if (diary.diaryId) {
-      fetchComments(true);
-    }
-  }, [diary.diaryId]);
 
   useEffect(() => {
     onCommentViewToggle?.(isCommentViewOpen);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [isCommentViewOpen]);
-
-  useEffect(() => {
-    if (scrollToCommentId) {
-      const element = document.getElementById(`comment-${scrollToCommentId}`);
-      if (element) {
-        element.scrollIntoView({ behavior: 'smooth', block: 'center' });
-      }
-      setScrollToCommentId(null);
-    }
-  }, [scrollToCommentId, comments, commentReplies, isCommentViewOpen]);
-
-  useEffect(() => {
-    const handlePopState = () => {
-      setIsCommentViewOpen(false);
-    };
-
-    if (isCommentViewOpen) {
-      window.history.pushState({ commentView: 'open' }, '');
-      window.addEventListener('popstate', handlePopState);
-    }
-
-    return () => {
-      window.removeEventListener('popstate', handlePopState);
-    };
-  }, [isCommentViewOpen]);
-
-  const handleSetReplyTo = (comment: Comment) => {
-    setReplyTo(comment);
-    setNewComment(`@${comment.nickname} `);
-    inputRef.current?.focus();
-  };
-
-  const cancelReply = () => {
-    setReplyTo(null);
-    setNewComment('');
-  };
-
-  const handleStartEdit = (comment: Comment) => {
-    setEditingComment(comment);
-    setReplyTo(null);
-    setNewComment(comment.content);
-    inputRef.current?.focus();
-  };
-
-  const handleCancelEdit = () => {
-    setEditingComment(null);
-    setNewComment('');
-  };
-
-  const handleSubmitComment = async () => {
-    if (editingComment) {
-      await handleUpdateComment(editingComment.id, newComment);
-      setEditingComment(null);
-      setNewComment('');
-    } else {
-      await handleCreateComment();
-    }
-  };
-
-  const handleCreateComment = async () => {
-    if (!isLoggedIn || !user || !newComment.trim()) return;
-
-    setIsSubmitting(true);
-    const tempId = Date.now();
-    const isReply = replyTo !== null;
-    const parentId = isReply && replyTo ? replyTo.id : undefined;
-
-    const contentToSend =
-      isReply && replyTo
-        ? newComment.replace(`@${replyTo.nickname} `, '')
-        : newComment;
-
-    const optimisticComment: Comment = {
-      id: tempId,
-      diaryId: diary.diaryId,
-      userId: String(user.id),
-      nickname: user.nickname || '사용자',
-      avatar: user.avatar || `${serverUrl}/globe.svg`,
-      content: contentToSend.trim(),
-      parentId: parentId || null,
-      createdAt: new Date().toISOString(),
-      updatedAt: new Date().toISOString(),
-      replyCount: 0,
-    };
-
-    setTotalComments(prev => prev + 1);
-    if (isReply && parentId) {
-      setComments(prev =>
-        prev.map(c =>
-          c.id === parentId ? { ...c, replyCount: c.replyCount + 1 } : c,
-        ),
-      );
-      const parentState = commentReplies[parentId] || {
-        list: [],
-        page: 0,
-        hasMore: true,
-        isLoading: false,
-        isShown: true,
-      };
-      setCommentReplies(prev => ({
-        ...prev,
-        [parentId]: {
-          ...parentState,
-          list: [...parentState.list, optimisticComment],
-          isShown: true,
-        },
-      }));
-    } else {
-      setComments(prev => [...prev, optimisticComment]);
-    }
-    setScrollToCommentId(tempId);
-    cancelReply();
-
-    try {
-      const newCommentData = await createComment({
-        diaryId: diary.diaryId,
-        content: contentToSend.trim(),
-        parentId,
-      });
-
-      const finalComment = { ...optimisticComment, ...newCommentData };
-
-      if (isReply && parentId) {
-        setCommentReplies(prev => {
-          const newReplies = prev[parentId].list.map(c =>
-            c.id === tempId ? finalComment : c,
-          );
-          return { ...prev, [parentId]: { ...prev[parentId], list: newReplies } };
-        });
-      } else {
-        setComments(prev =>
-          prev.map(c => (c.id === tempId ? finalComment : c)),
-        );
-      }
-    } catch (error) {
-      console.error('댓글 작성 실패:', error);
-      alert('댓글 작성에 실패했습니다.');
-      // Revert optimistic update
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
-
-  const handleDeleteComment = async (
-    commentId: number,
-    parentId: number | null,
-  ) => {
-    // Optimistic update
-    if (parentId) {
-      setCommentReplies(prev => ({
-        ...prev,
-        [parentId]: {
-          ...prev[parentId],
-          list: prev[parentId].list.filter(c => c.id !== commentId),
-        },
-      }));
-      setComments(prev =>
-        prev.map(c =>
-          c.id === parentId ? { ...c, replyCount: c.replyCount - 1 } : c,
-        ),
-      );
-    } else {
-      setComments(prev => prev.filter(c => c.id !== commentId));
-    }
-    setTotalComments(prev => prev - 1);
-
-    try {
-      await deleteComment(commentId);
-    } catch (error) {
-      console.error('댓글 삭제 실패:', error);
-      alert('댓글 삭제에 실패했습니다.');
-      // Revert optimistic update (optional, for simplicity we'll just log)
-    }
-  };
-
-  const handleUpdateComment = async (commentId: number, content: string) => {
-    if (!content.trim()) return;
-
-    let originalComments = comments;
-    let originalReplies = commentReplies;
-
-    const updateInComments = (list: Comment[]): Comment[] =>
-      list.map(c => (c.id === commentId ? { ...c, content } : c));
-
-    setComments(prev => updateInComments(prev));
-
-    setCommentReplies(prev => {
-      const newReplies = { ...prev };
-      for (const parentId in newReplies) {
-        newReplies[parentId] = {
-          ...newReplies[parentId],
-          list: updateInComments(newReplies[parentId].list),
-        };
-      }
-      return newReplies;
-    });
-
-    try {
-      await updateComment(commentId, content);
-    } catch (error) {
-      console.error('댓글 수정 실패:', error);
-      alert('댓글 수정에 실패했습니다.');
-      setComments(originalComments);
-      setCommentReplies(originalReplies);
-    }
-  };
 
   const handlePrevImage = () => {
     if (diary.imgUrls && diary.imgUrls.length > 0) {
@@ -411,15 +84,6 @@ const DiaryStoryModal = ({
   const handleDragEnd = (event: MouseEvent | TouchEvent | PointerEvent, info: PanInfo) => {
     if (info.offset.y < -50 && info.velocity.y < -200) {
       setIsCommentViewOpen(true);
-    }
-  };
-
-  const handleCommentViewDragEnd = (
-    event: MouseEvent | TouchEvent | PointerEvent,
-    info: PanInfo,
-  ) => {
-    if (info.offset.y > 50 && info.velocity.y > 200) {
-      setIsCommentViewOpen(false);
     }
   };
 
@@ -528,104 +192,22 @@ const DiaryStoryModal = ({
         </motion.div>
       )}
 
-      {/* Comments View */}
+      {/* Comments View using StoryCommentModal */}
       <AnimatePresence>
         {isCommentViewOpen && (
-          <motion.div
-            className="absolute bottom-0 left-0 right-0 z-30 flex h-[85%] cursor-grab flex-col rounded-t-2xl bg-white shadow-lg dark:bg-gray-800"
-            initial={{ y: '100%' }}
-            animate={{ y: '0%' }}
-            exit={{ y: '100%' }}
-            transition={{ type: 'spring', stiffness: 300, damping: 30 }}
-            drag="y"
-            dragConstraints={{ top: 0, bottom: 0 }}
-            dragElastic={{ top: 0, bottom: 0.5 }}
-            onDragEnd={handleCommentViewDragEnd}
-          >
-            <div
-              className="flex-shrink-0 cursor-pointer border-b border-gray-200 p-4 text-center dark:border-gray-700"
-              onClick={() => setIsCommentViewOpen(false)}
-            >
-              <span className="inline-block h-1.5 w-10 rounded-full bg-gray-300 dark:bg-gray-600" />
-            </div>
-
-            <div className="flex-grow space-y-4 overflow-y-auto p-4 no-scrollbar">
-              {/* Diary Content */}
-              <div className="flex items-start" onClick={handleProfileClick}>
-                <img
-                  src={diary.avatar || DEFAULT_AVATAR}
-                  alt={diary.nickname}
-                  width={32}
-                  height={32}
-                  className="mr-3 mt-1 cursor-pointer rounded-full"
-                  onError={handleAvatarError}
-                />
-                <div>
-                  <p className="whitespace-pre-wrap text-sm dark:text-white">
-                    <span className="cursor-pointer font-bold">{diary.nickname}</span>{' '}
-                    {diary.content}
-                  </p>
-                  <p className="mt-1 text-xs text-gray-400 dark:text-gray-500">
-                    {formatTimeAgo(diary.createdAt)}
-                  </p>
-                </div>
-              </div>
-              
-              {/* Comments */}
-              {comments.map(comment => (
-                <CommentItem
-                  key={comment.id}
-                  comment={comment}
-                  diaryId={diary.diaryId}
-                  onSetReplyTo={handleSetReplyTo}
-                  replies={commentReplies[comment.id]?.list || []}
-                  replyState={commentReplies[comment.id]}
-                  onToggleReplies={() => handleToggleReplies(comment)}
-                  onFetchMoreReplies={() => handleFetchReplies(comment.id)}
-                  onDeleteComment={handleDeleteComment}
-                  onStartEdit={handleStartEdit}
-                />
-              ))}
-              {comments.length === 0 && !isLoadingComments && (
-                <div className="py-4 text-center text-gray-500">
-                  아직 댓글이 없습니다.
-                </div>
-              )}
-              {isLoadingComments && <div className="py-2 text-center text-gray-500">댓글을 불러오는 중...</div>}
-              {!isLoadingComments && hasMore && (
-                <div className="py-2 text-center">
-                  <button onClick={() => fetchComments()} className="text-sm font-semibold text-gray-500 hover:text-gray-800 dark:text-gray-400 dark:hover:text-gray-200">
-                    다음 댓글 더 보기
-                  </button>
-                </div>
-              )}
-            </div>
-
-            {/* Comment Input */}
-            <div className="border-t border-gray-200 bg-white p-4 dark:border-gray-700 dark:bg-gray-800">
-              <CommentInput
-                inputRef={inputRef as React.RefObject<HTMLInputElement>}
-                onSubmit={handleSubmitComment}
-                placeholder={
-                  editingComment
-                    ? '댓글 수정...'
-                    : replyTo
-                      ? `@${replyTo.nickname}님에게 답글 남기기`
-                      : '댓글 달기...'
-                }
-                value={newComment}
-                onChange={setNewComment}
-                isSubmitting={isSubmitting}
-                onCancel={
-                  editingComment
-                    ? handleCancelEdit
-                    : replyTo
-                      ? cancelReply
-                      : undefined
-                }
-              />
-            </div>
-          </motion.div>
+          <StoryCommentModal
+            diaryId={diary.diaryId}
+            initialCommentCount={totalComments}
+            onClose={handleCloseCommentModal}
+            onUpdateCommentCount={handleUpdateCommentCount}
+            diaryContent={{
+              nickname: diary.nickname,
+              avatar: diary.avatar,
+              content: diary.content,
+              createdAt: diary.createdAt,
+              userId: diary.userId,
+            }}
+          />
         )}
       </AnimatePresence>
     </motion.div>
